UI projects updater widgets now starting socket.io connection on page load
